// src/pages/Register/components/RegisterForm/types.ts
export type RegisterFormData = {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    confirmPassword: string;
    phoneNumber: string;
    addressLine1: string;
    addressLine2: string;
    state: string;
    locality: string;
    postalCode: string;
    country: string;
}